backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.2
python-dotenv==1.0.1
pydantic==2.7.1
apscheduler==3.10.4
chromadb==0.5.0
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.2